Change EffortDisplay superclass from MessageFilterDisplay to RosTopicDisplay to avoid dropping messages with empty frame_id.
Problem: EffortDisplay does not react on JointState messages with emty frame_id.
Also the following warning is displayed:

  [rviz2]: Message Filter dropping message: frame '' at time ..... for reason
  'the frame id of the message is empty'

Reason: MessageFilterDisplay uses tf2_ros::MessageFilter which checks
that frame_id is not empty. This mechanism is designed to perform message
frame change "on-fly" but it is nao applicable to JointState messages.

Solution: Use RosTopicDisplay instead of MessageFilterDisplay. Use type erased signal/slot
pair to ensure messages are processed in the main thread.